+++ src/version/index.ts
@@ -1,4 +1,4 @@
-export const VERSION = '4.0.33';
+export const VERSION = '5.0.21';
 
 export function getVersion(): string {
   return VERSION;